requirements.txt
asgiref==3.7.2
chardet==5.2.0
config==0.5.1
decouple==0.0.7
Django==4.2.8
djangorestframework==3.14.0
gunicorn==21.2.0
install==1.3.5

packaging==23.2

psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
unicodecsv==0.14.1
